requirements.txt
asgiref==3.2.7
backcall==0.2.0
colorama==0.4.3
cycler==0.10.0
decorator==4.4.2
Django==3.0.6
django-autocomplete-light==3.5.1
django-crispy-forms==1.9.1
django-filter==2.2.0
django-widget-tweaks==1.4.8
image==1.5.32
ipython==7.17.0
ipython-genutils==0.2.0
jedi==0.17.2
kiwisolver==1.2.0
matplotlib==3.2.1
numpy==1.18.4
pandas==1.0.4
parso==0.7.1
pickleshare==0.7.5
Pillow==7.1.2
prompt-toolkit==3.0.6
Pygments==2.6.1
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.1
requests==2.23.0
six==1.15.0
sqlparse==0.3.1
traitlets==4.3.3
urllib3==1.25.9
wcwidth==0.2.5
django-tables2==2.3.3
sorl-thumbnail==12.7.0
django-imagekit==4.0.2
pylint-django==2.3.0